Add automatic deployment and documentation
